refactor: 🔧 update workspace response types and improve import paths
