--- client/apps/web/src/workspace/domain/usecases/index.ts
@@ -1,6 +1,0 @@
-/**
- * Workspace use cases for business logic operations.
- */
-
-export { GetWorkspaceById } from "./GetWorkspaceById";
-export { ListWorkspaces } from "./ListWorkspaces";
